fix(cad-viewer): move MlLayerTable types out of Vue SFC for TypeDoc
