+++ packages/cad-viewer/src/component/common/index.ts
@@ -11,11 +11,11 @@ export { default as MlFieldsetGroup } from './MlFieldsetGroup.vue'
 export { default as MlHatchPatternPanel } from './MlHatchPatternPanel.vue'
 export { default as MlLineWeightSelect } from './MlLineWeightSelect.vue'
 export { default as MlLayerSelect } from './MlLayerSelect.vue'
-export { default as MlLayerTable } from './MlLayerTable.vue'
 export type {
   MlLayerTableChangeField,
   MlLayerTableRow
-} from './MlLayerTable.vue'
+} from './MlLayerTable'
+export { default as MlLayerTable } from './MlLayerTable.vue'
 export { default as MlFontFileReader } from './MlFontFileReader.vue'
 export { default as MlLineTypeSelect } from './MlLineTypeSelect.vue'
 export { default as MlHatchPatternDropdown } from './MlHatchPatternDropdown.vue'
